Use webpack for bundling dependencies and frontend code.
Pushing this out to git to save the progress that I've been working on. I have no idea anymore if this is a good idea or a bad idea but I've already invested a few hours into this and deleting it all would be stupid. Still have to cleanup the code that has been shoved into modules and figure out the rest of the breaking changes made with all these changes.