Delay spamless task calls so they're less likely to happen before commit.
This isn't 100% guaranteed to work. The only thing is is calling db.commit()
before celery.send_task(), but I don't really want to commit in send_message()
because that expires everything in the database session so we'd end up issuing
extra SELECT queries for any database objects we access later.

To do:
* Check the million places where we call send_message() to see what they do
  afterwards.
* Consider setting expire_on_commit to False so we can read objects after a
  commit. Could be dangerous if we try to modify them though.